requirements.txt
absl-py
torch
torchvision
pillow
flask
psycopg2-binary
werkzeug
flask-cors
requests
numpy